BigInteger xatosi tuzatildi

+++ app/models/order.py
@@ -2,13 +2,15 @@ from datetime import datetime
 from sqlalchemy import DateTime, ForeignKey, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.database.base import Base
+from sqlalchemy import DateTime, ForeignKey, Integer, BigInteger, String, func
+
 
 
 class Order(Base):
     __tablename__ = "orders"
 
     id: Mapped[int] = mapped_column(primary_key=True)
-    user_id: Mapped[int] = mapped_column(Integer, nullable=False, index=True)
+    user_id: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
     shop_id: Mapped[int] = mapped_column(ForeignKey("shops.id"), nullable=False)
     total_price: Mapped[int] = mapped_column(Integer, nullable=False)
     status: Mapped[str] = mapped_column(String(30), default="new", nullable=False)
